Fix TypeError in create_match_filter for Composite Keys (#1693)
**Old Code Behavior:**

Even if there's only one such condition, the code wraps it in an Or()
operator.
But Or() is meant to combine two or more conditions (like “condition A
OR condition B”).
If you give it only one condition, it complains because it expects a
second condition.

**New Code Behavior:**

The new change checks how many conditions you have.
If there's only one condition, it simply returns that condition.
If there are more than one, it uses Or() to combine them.

---------

Co-authored-by: Fokko Driesprong <fokko@apache.org>