Remove parsing table name in `row_filter` (#1689)
This PR deprecates one of the three items that were planned for the
0.9.0 release.

All items marked for removal:
- [x]  Table name reference in scan expression 

<https://github.com/apache/iceberg-python/blob/efc8b5ac0f16717f776e034ecf9a9e9bdabd8424/pyiceberg/expressions/parser.py#L95>

- [x]  REST catalog client AUTH_URL (#1691)

<https://github.com/apache/iceberg-python/blob/efc8b5ac0f16717f776e034ecf9a9e9bdabd8424/pyiceberg/catalog/rest.py#L324>

- [x] botocore session (#1692)

<https://github.com/apache/iceberg-python/blob/efc8b5ac0f16717f776e034ecf9a9e9bdabd8424/pyiceberg/catalog/__init__.py#L790>

Currently there are three items marked for release. However, based on
the ongoing
[discussion](https://lists.apache.org/thread/rr8lcf96jl6079dz6vfkwr5spbvlxzpm),
it appears that the other two items. have not yet been replaced with a
proper solution. As a result, this PR only addresses the deprecation of
`Table name reference in scan expression` while we await further
resolution on the others.