CI: Use Java 1.9.0-SNAPSHOT for testing (#1899)
With https://github.com/apache/iceberg/pull/12593/ Row lineage is
required and Spark uses Java 1.8.0 that does not write the required
`first-row-id` field in the snapshot.

Therefore, I think it would be good to just use the SNAPSHOT for now
until it gets released. We can easily revert the PR once that's done.

Closes #1898

<!--
Thanks for opening a pull request!
-->

<!-- In the case this PR will resolve an issue, please replace
${GITHUB_ISSUE_ID} below with the actual Github issue id. -->
<!-- Closes #${GITHUB_ISSUE_ID} -->

# Rationale for this change

# Are these changes tested?

# Are there any user-facing changes?

<!-- In the case of user-facing changes, please add the changelog label.
-->