Make `Record` purely position based (#1768)
This aligns the implementation with Java.

We had the keywords there mostly for the tests, but they should not be
used, and it seems like that's already the case :'( I was undecided if
the costs of this PR (all the changes), are worth it, but I see more PRs
using the Record in a bad way (example
https://github.com/apache/iceberg-python/pull/1743) that might lead to
very subtle bugs where the position might sometime change based on the
ordering of the dict.

Blocked by https://github.com/Eventual-Inc/Daft/pull/3917