add PARTITION_SUMMARY_PROP (#2202)
<!--
Thanks for opening a pull request!
-->

<!-- In the case this PR will resolve an issue, please replace
${GITHUB_ISSUE_ID} below with the actual Github issue id. -->
<!-- Closes #${GITHUB_ISSUE_ID} -->

# Rationale for this change

Adding the snapshot property to keep parity with
[Java](https://github.com/apache/iceberg/commits/main/core/src/main/java/org/apache/iceberg/SnapshotSummary.java)

# Are these changes tested?

Yes I modified two of the existing tests to add this property and added
an additional two for cases where the table is unchanged or
unpartitioned.

# Are there any user-facing changes?

Yes, this will add a property to the Snapshot Summaries when partitions
are changed.

<!-- In the case of user-facing changes, please add the changelog label.
-->

---------

Co-authored-by: Giorgio Trettenero <gtret@netflix.com>
Co-authored-by: Kevin Liu <kevin.jq.liu@gmail.com>