Add import check for optional dependency on pyiceberg_core (#2221)
Added a NotInstalledException check when pyiceberg_core is imported for
pyarrow_transforms This will raise a helpful error message when endusers
try to use methods that depend on pyiceberg_core but haven't installed
the optional dependency

<!--
Thanks for opening a pull request!
-->

Closes #1987 

# Rationale for this change
If an enduser hasn't installed the `pyiceberg_core` optional dependency,
using pyarrow transforms will crash with an unhelpful error. This PR
gives the enduser a nicer error message that informs them to install the
optional dependency

# Are these changes tested?
Yes, a test was added to verify the behaviour

# Are there any user-facing changes?

<!-- In the case of user-facing changes, please add the changelog label.
-->

---------

Co-authored-by: Anders Bogsnes <andersbognes@gmail.com>