revert avro timestamp-millis mapping (#2223)
<!--
Thanks for opening a pull request!
-->

<!-- In the case this PR will resolve an issue, please replace
${GITHUB_ISSUE_ID} below with the actual Github issue id. -->
<!-- Closes #${GITHUB_ISSUE_ID} -->

# Rationale for this change
Follow up to #2173, the conversion as implemented is not safe since
iceberg does not natively have `millis` precision. Timestamp values are
implicitly in `micros` precision
This is confirmed by
https://github.com/apache/iceberg-python/pull/2173#discussion_r2209756001

Let's revert for now and follow up to figure out the right way to
support `millis` precision.

# Are these changes tested?

# Are there any user-facing changes?

<!-- In the case of user-facing changes, please add the changelog label.
-->